Refactor providers library and remove unused files
