--- libs/providers/README.md
@@ -1,11 +1,0 @@
-# providers
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build providers` to build the library.
-
-## Running unit tests
-
-Run `nx test providers` to execute the unit tests via [Jest](https://jestjs.io).
