--- libs/providers/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@autonolas-frontend-mono/providers",
-  "version": "0.0.1",
-  "dependencies": {
-    "tslib": "^2.3.0"
-  },
-  "type": "commonjs",
-  "main": "./src/index.js",
-  "typings": "./src/index.d.ts"
-}
