--- libs/providers/project.json
@@ -1,30 +1,0 @@
-{
-  "name": "providers",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/providers/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/libs/providers",
-        "main": "libs/providers/src/index.ts",
-        "tsConfig": "libs/providers/tsconfig.lib.json",
-        "assets": ["libs/providers/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": ["{options.outputFile}"]
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/providers/jest.config.ts"
-      }
-    }
-  },
-  "tags": []
-}
