+++ libs/ui-theme/src/lib/ThemeConfig.tsx
@@ -1,5 +1,5 @@
-import React, { ReactNode } from 'react';
-import { ConfigProvider, ThemeConfig } from 'antd';
+import React, { useEffect, useRef } from 'react';
+import { ConfigProvider, ConfigProviderProps, ThemeConfig } from 'antd';
 import { COLOR } from './ui-theme';
 
 export const THEME_CONFIG: ThemeConfig = {
@@ -36,10 +36,23 @@ export const THEME_CONFIG: ThemeConfig = {
   },
 };
 
+/**
+ * Use this provider in place of the `ConfigProvider` from `antd` to
+ * ensure that the theme is only applied once the component has mounted.
+ * @param props The props to pass to the `ConfigProvider`.
+ */
 export const AutonolasThemeProvider = ({
-  children,
-}: {
-  children: ReactNode;
-}) => {
-  return <ConfigProvider theme={THEME_CONFIG}>{children}</ConfigProvider>;
+  theme = THEME_CONFIG,
+  ...rest
+}: ConfigProviderProps) => {
+  const isMounted = useRef(false);
+
+  useEffect(() => {
+    isMounted.current = true;
+    return () => {
+      isMounted.current = false;
+    };
+  }, []);
+
+  return isMounted.current ? <ConfigProvider theme={theme} {...rest} /> : null;
 };
